requirements.txt
numpy==1.26.4
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
torchmetrics
pesq
librosa
tqdm
matplotlib
pandas
wandb
comet_ml
hydra-core

black
isort
pre-commit
